==> 1.ipynb <==
# converting mp3 to wav.

## Bash solution:

!rm sample.wav; ffmpeg -i sample.mp3 -ar 16000 sample.wav

## Python equivalent:

from pydub import AudioSegment

AudioSegment.from_mp3("sample.mp3").export("sample.wav", format="wav", parameters=["-ar", "16000"])

## Proposed but untested implementation in Transformers wav2vec pipeline:


import tempfile
from pydub import AudioSegment
def load_audio(path):
    with tempfile.NamedTemporaryFile() as f:
        AudioSegment.from_mp3(path).export(
                f.name,
                format="wav", 
                parameters=["-ar", "16000"]
                )
        return datasets.Audio(sampling_rate=16000).decode_example(f.name)

%%timeit
AudioSegment.from_mp3("sample.mp3").export("sample.wav", format="wav", parameters=["-ar", "16000"])

s = AudioSegment.from_mp3("sample.mp3")
s.__len__() # Lenths in ms

from pint import UnitRegistry
u = UnitRegistry()

result = 1976 * u.hour * 84.2*u("millisecond") / (8.7 * u.s)
result.to(u.hour)
