==> 2_sampling.ipynb <==
# Objective:

```
Produce a random selection of 1000 20sec segment pairs (text,mp3) from the data below, to check their quality. Important is 1. these are random from the whose sample, 2. it is easy to inspect the transcript and find the corresponding mp3 file, I suggest you just use the “alphabet ordering” trick, so mp3 files called “samplexxx”, xxx ranging from 000 to 999. One text file can have these file names and transcripts in file-name-alphabetical order.

20sec segments
/home/korzinek/kaldi/exp/ali_all/ali.json

Mapping of recoXXXX ids to actual file wav names
/home/korzinek/kaldi/exp/ali_all/wav.scp

Symbolic links to wavs
/home/nikola/projects/parlaspeech/wav/
```


# File management

links_to_wavs = "/home/nikola/projects/parlaspeech/wav/"
mappingsfile = "/home/korzinek/kaldi/exp/ali_all/wav.scp"
segmentsfile = "/home/korzinek/kaldi/exp/ali_all/asr.json"

import json
with open(segmentsfile, "r") as file:
    contents = []
    for line in file.readlines():
        contents.append(json.loads(line))


import pandas as pd
segments = pd.DataFrame(contents)
segments.head()

reconame, longname = [],[]
with open(mappingsfile) as f:
    for line in f.readlines():
        reconame.append(line[0:7])
        longname.append(line[8:-1])

mappings = pd.DataFrame(data={"reconame": reconame, "longname": longname})
mappings.shape

merged = segments.merge(mappings, how="left", left_on="file", right_on="longname").drop(columns="longname")
merged.head(3)

Remarks:
* I have the long filename and the recoNNN mapping merged. Not sure why I need the latter tho.
* Wav files are present at `links_to_wavs + long filename`.

Plan:
* Generate a sample of 1000 rows from the merged DF.
* Based on the info I have in the merged dataframe I will generate transcript from `words` column
* I will trim the audio file to the segment borders.
* Save the transcript, save the audio.

sampledf = merged.sample(1000).reset_index(drop=True)
sampledf["transcript"] = sampledf.words.str.join(" ")
sampledf["asr_transcript"] = sampledf.asr.str.join(" ")
sampledf["samplename"] = ""
sampledf["samplepath"] = ""

sampledf.file[0]

from pydub import AudioSegment
import os
output_dir = "/home/rupnik/macocu/task8/sample"
for i, row in sampledf.iterrows():
    print(f"i: {i:0004}", end="\r")
    wavfile = row["file"]
    transcript = row["transcript"]
    reconame = row["reconame"]
    segment_start_ms = int(row["start"] * 1000)
    segment_end_ms = int(row["end"] * 1000)
    try:
        audio = AudioSegment.from_wav(
            links_to_wavs  + wavfile
            )
        segment = audio[segment_start_ms:segment_end_ms]
        outpath = os.path.join(
                output_dir, 
                f"sample{i:003}.mp3")
        segment.export(
            outpath, 
            format="mp3")
        sampledf.loc[i, "samplepath"] = outpath
        sampledf.loc[i, "samplename"] = f"sample{i:003}.mp3"
    except FileNotFoundError:
        continue



finished = !ls /home/rupnik/macocu/task8/sample
sampledf["finished"] = sampledf.samplename.isin(finished)
sampledf.head(10)#to_excel("transcripts.xls")

sampledf.loc[sampledf.samplename != "", :].to_csv("transcripts.csv")

from pydub import AudioSegment
import os
output_dir = "/home/rupnik/macocu/task8/sample"
for i, row in sampledf.iterrows():
    if row["samplename"] != "":
        continue
    print(f"i: {i:0004}", end="\r")
    wavfile = row["file"]
    transcript = row["transcript"]
    reconame = row["reconame"]
    segment_start_ms = int(row["start"] * 1000)
    segment_end_ms = int(row["end"] * 1000)
    try:
        audio = AudioSegment.from_wav(
            links_to_wavs  + wavfile.replace("-", "- ")
            )
        segment = audio[segment_start_ms:segment_end_ms]
        outpath = os.path.join(
                output_dir, 
                f"sample{i:003}.mp3")
        segment.export(
            outpath, 
            format="mp3")
        sampledf.loc[i, "samplepath"] = outpath
        sampledf.loc[i, "samplename"] = f"sample{i:003}.mp3"
    except FileNotFoundError:
        continue




sampledf.loc[sampledf.samplename != "", :].to_csv("transcripts.csv")

available_files = os.listdir(links_to_wavs)

from pydub import AudioSegment
import os
output_dir = "/home/rupnik/macocu/task8/sample"
for i, row in sampledf.iterrows():
    if row["samplename"] != "":
        continue
    print(f"i: {i:0004}", end="\r")
    wavfile = row["file"]
    transcript = row["transcript"]
    reconame = row["reconame"]
    segment_start_ms = int(row["start"] * 1000)
    segment_end_ms = int(row["end"] * 1000)

    if wavfile not in available_files:
        split_filename = wavfile.split()
        for available_file in available_files:
            if split_filename == available_file.split():
                wavfile = available_file
    try:
        audio = AudioSegment.from_wav(
            links_to_wavs  + wavfile
            )
        segment = audio[segment_start_ms:segment_end_ms]
        outpath = os.path.join(
                output_dir, 
                f"sample{i:003}.mp3")
        segment.export(
            outpath, 
            format="mp3")
        sampledf.loc[i, "samplepath"] = outpath
        sampledf.loc[i, "samplename"] = f"sample{i:003}.mp3"
    except FileNotFoundError:
        continue

failed_condition = sampledf.samplename==""
failed_files = set(sampledf.loc[failed_condition, "file"].unique())
ok_files = set(sampledf.loc[~failed_condition, "file"].unique())

ok_files.intersection(failed_files)

failed_files

mappings
