==> 4_preparing_to_move_data.ipynb <==
import pandas as pd
df = pd.read_csv("parlaspeech.csv")
df.columns

df_pruned = df.loc[:, ['file', 'start', 'end',
     'true_file', 'reconame', 'yt_hash', 'hashname',
     'was_in_samples', 'human_transcript', 'asr_transcript', 'sim',]]

df_pruned.to_csv("parlaspeech_pruned.csv", index=False)

df_pruned.to_json("parlaspeech_pruned.json", orient="records", lines=True)

estimated_size = df_pruned.shape[0]/320174 * 321
estimated_size

c_to_drop = df_pruned.was_in_samples | df_pruned.sim<0.8
df_pruned = df_pruned.drop(index=df_pruned.index[c_to_drop])
available = 232141524 
corpus_size = 450725396

train_fraction = available/corpus_size
train_fraction

Zaokrožim na 0.5

keep_df = df_pruned.sample(frac=0.5, random_state=42)
keep_df.to_csv("keep_pruned.csv", index=False)

from concurrent.futures import ProcessPoolExecutor
from shutil import copyfile
import os
def copy(file):
    if os.path.exists(os.path.join("/home/rupnik/macocu/task8/transfer", file)):
        return

    copyfile(os.path.join("/home/rupnik/macocu/task8/data_16000", file),
            os.path.join("/home/rupnik/macocu/task8/transfer", file)
            )
with ProcessPoolExecutor(max_workers=100) as executor:
    executor.map(copy, keep_df.hashname.values)
