==> 5_vocab_preparation.ipynb <==
import pandas as pd
df = pd.read_csv("keep_pruned.csv")
human_transcripts = df.human_transcript.values
asr_transcripts = df.asr_transcript.values

chars = set()
for line in human_transcripts:
    try:
        chars = chars.union(set(line))
    except:
        continue
for line in asr_transcripts:
    try:
        chars = chars.union(set(line))
    except:
        continue
chars

len(chars)

chars = chars.union({"[PAD]", "[UNK]"})


i = 0
d = dict()
for c in sorted(chars):
    d[c] = i
    i+=1
with open("vocab.json", "w") as f:
    import json
    json.dump(d, f)
