==> 7_correcting_the_dataset_for_200h.ipynb <==
import pandas as pd
df = pd.read_csv("keep_pruned.csv")
df["duration"] = df.end - df.start
df["cumulative_duration"] = df.duration.cumsum()
df.head()

c = df.cumulative_duration < 200*3600
df.loc[c, :].to_csv("keep_pruned.csv")

from concurrent.futures import ProcessPoolExecutor
import os
def transform(file):
    if os.path.exists(f"data_16000_mono/{file}"):
        return
    os.system(f"ffmpeg -loglevel panic -y -i data_16000/{file} -ac 1 data_16000_mono/{file}")
with ProcessPoolExecutor(max_workers=32) as executor:
    executor.map(
        transform, 
        os.listdir("/home/rupnik/macocu/task8/data_16000/")
        )

df.loc[c, :].tail()

200*3600


import subprocess
direc = "data_16000_mono/"
files = os.listdir(direc)
r = subprocess.check_output([f"soxi -r {os.path.join(direc, files[0])}"], shell=True)
set([subprocess.check_output(["soxi", "-r", os.path.join(direc, file)], shell=True) for file in files])

set([subprocess.check_output(["soxi", "-c", os.path.join(direc, file)], shell=True) for file in files])
