==> 8_adding_unnormalised_transcriptions.ipynb <==
import pandas as pd
unnorm_path = "/home/korzinek/kaldi/exp/ali_all/unnorm.json"
import json
contents = list()
with open(unnorm_path, "r") as file:
    for line in file.readlines():
        contents.append(json.loads(line))
import pandas as pd
unnorm = pd.DataFrame(contents)

df = pd.read_csv("parlaspeech.csv")
df.head(2)

unnorm = unnorm.rename(columns={
    "words": "words_unnormalized",
    "word_start_times": "word_start_times_unnormalized",
    "word_end_times": "word_end_times_unnormalized"
})
unnorm.head(2)

df.shape, unnorm.shape

df.columns

unnorm["key"] = unnorm.apply(lambda row: row.file+str(row.start)+str(row.end), axis = 1)

unnorm.key[unnorm.key.isin(df.key.unique())].shape[0] / df.shape[0]

We cover 97% of the instances. Let's merge the two resources.

unnorm.columns

merged = df.merge(unnorm.loc[:, ['words_unnormalized',
       'word_start_times_unnormalized', 'word_end_times_unnormalized', 'key']], left_on="key", right_on="key", how="left")

merged.head(2)

merged.columns

merged["unnormalized_transcript"] = merged.words_unnormalized.str.join(" ")

merged.head(2)

merged.words_unnormalized.isna().sum()

merged.to_csv("parlaspeech.csv", index=False)
merged.to_json("parlaspeech.json", orient="records", lines=True)

condition = ~merged.words_unnormalized.isna()
selection = merged.loc[condition,:]
selection.columns

selection.to_csv("selection.csv", index=False)
selection.to_json("selection.json", orient="records", lines=True)

selection.hashname[0:5]

from concurrent.futures import ProcessPoolExecutor
import os
os.system("mkdir selection_16000_mono")
def copy(file):
    os.system(f"cp data_16000_mono/{file} selection_16000_mono/{file}")
with ProcessPoolExecutor(max_workers=32) as executor:
    executor.map(
        copy,
        selection.hashname)


in_selection = set(selection.hashname.values)
in_transfer = set(os.listdir("transfer/"))


len(in_transfer.difference(in_selection))

in_transfer

len(in_transfer)

len(in_selection)
