==> 9_inspecting_Nikola's_DS.ipynb <==
f = "/home/nikola/projects/parlaspeech/ParlaSpeech-HR.v1.0.jsonl"

import pandas as pd
n = pd.read_json(f, orient="records", lines=True)

n.columns

n.head(2)

ps = pd.read_csv("parlaspeech.csv")

ps_files = set(ps.hashname.values.tolist())
n_files = set(n.path.values.tolist())

len(ps_files.difference(n_files))

import os
actual_files = set(os.listdir("selection_16000_mono/"))
len(actual_files.difference(n_files))

s = pd.read_csv("selection.csv")

selection_files = set(s.hashname.values.tolist())
len(selection_files.difference(n_files))

(s.sim < 0.8).sum()

n.shape

s.shape

s["is_in_n"] = s.hashname.isin(n.path)

s.is_in_n.sum()

s.loc[s.is_in_n, :].drop(columns="is_in_n").to_csv("selection.csv", index=False)
s.loc[s.is_in_n, :].drop(columns="is_in_n").to_json("selection.json", orient = "records", lines=True)



from concurrent.futures import ProcessPoolExecutor
import os

def delete(file):
    os.system(f"rm selection_16000_mono/{file}")
with ProcessPoolExecutor(max_workers=32) as executor:
    executor.map(
        delete,
        s.loc[~s.is_in_n, "hashname"])
