==> 10_preparing_new_transfer_dir.ipynb <==
import pandas as pd
sel = pd.read_csv("selection.csv")
sel.head(2)

sel["duration"] = sel.end - sel.start
sel["cumulative_duration"] = sel.duration.cumsum()
sel.loc[sel.cumulative_duration < 200 * 3600, :].tail(2)

sel.iloc[0:44562+1, :].to_csv("transfer_10.csv", index=False)

sel_200h = sel.iloc[0:44562+1, :].copy()

sel_200h.hashname.head()

from concurrent.futures import ProcessPoolExecutor
import os
def copy(file):
    os.system(f"cp selection_16000_mono/{file} transfer_10/{file}")
with ProcessPoolExecutor(max_workers=32) as executor:
    executor.map(
        copy,
        sel_200h.hashname)
