==> 11_testing_data.ipynb <==
import pandas as pd
df = pd.read_csv("transfer_10.csv")
df.loc[:, ["hashname", "human_transcript"]].sample(10)

sample = [
"5Kq_FBO-iDk_6021.8-6039.35.wav",  "dfz6VqXRbmE_11159.0-11178.81.wav",
"5T4HXO1WsDU_7431.84-7451.82.wav",  "MJ9sB3Mu0Sw_13121.95-13136.8.wav",
"8LAI5kcBfVs_7658.83-7675.22.wav" , "wN5mbIwE-4U_9720.21-9735.96.wav",
"-aa4kot7dUU_6398.42-6415.91.wav" , "WoEEFk8lcDA_7909.43-7928.86.wav",
"bv67e1t69T8_12814.88-12823.3.wav" , "y5LrUr3SHIg_11332.37-11352.27.wav",
]
pd.set_option('display.max_colwidth', -1)  # or 199
df.loc[df.hashname.isin(sample), ["hashname", "human_transcript"]]
