==> 12_dataset_inspection.ipynb <==
import pandas as pd

df = pd.read_csv("transfer_10.csv")


import os


# Are all of the files sampled at 16kHz?

df.hashname[0]

import subprocess
subprocess.getoutput("soxi -r transfer_10/sPNuDebATRM_1853.82-1873.66.wav")

def get_samplerate(file):
    return subprocess.getoutput(f"soxi -r transfer_10/{file}")


from concurrent.futures import ProcessPoolExecutor
with ProcessPoolExecutor(max_workers=20) as executor:
    samplerates = executor.map(get_samplerate, df.hashname)
print("Detected samplerates in transfer_10 directory: ", set(samplerates))

def get_channels(file):
    return subprocess.getoutput(f"soxi -c transfer_10/{file}")


from concurrent.futures import ProcessPoolExecutor
with ProcessPoolExecutor(max_workers=20) as executor:
    channels = executor.map(get_channels, df.hashname)
print("Detected channels in transfer_10 directory: ", set(channels))

def get_bitrate(file):
    return subprocess.getoutput(f"soxi -B transfer_10/{file}")


from concurrent.futures import ProcessPoolExecutor
with ProcessPoolExecutor(max_workers=20) as executor:
    bitrate = executor.map(get_bitrate, df.hashname)
print("Detected bitrate in transfer_10 directory: ", set(bitrate))

def get_precision(file):
    return subprocess.getoutput(f"soxi -b transfer_10/{file}")


from concurrent.futures import ProcessPoolExecutor
with ProcessPoolExecutor(max_workers=20) as executor:
    precision = executor.map(get_precision, df.hashname)
print("Detected precision in transfer_10 directory: ", set(precision))
