==> 17_new_dataset.ipynb <==
import pandas as pd

df = pd.read_csv("transfer_10.csv").dropna()
df.shape

Let's remove instances that were in samples and instances where sim < 0.8:

c1 = ~df.was_in_samples
c2 = df.sim >= 0.8

df = df.loc[c1 & c2, :]


Let's prepare a subset of 25k instances and calculate the duration:

subset = df.iloc[:25500].reset_index(drop=True)
((subset.end - subset.start).cumsum()/3600).values[-1]

Ok, we have 114.7 hours of data. Let's randomize the row order instead of doing train-dev-test split:

subset = subset.sample(frac=1).reset_index(drop=True)

subset["split"] = ["train"] * 24500 + ["test"] * 500 + ["dev"] * 500

condition = subset.split == "test"
subsubset = subset.loc[condition, :]
(subsubset.end-subsubset.start).sum()/3600

train = subset.loc[
    subset.split=="train",  ['hashname',
                            'human_transcript',
                            'unnormalized_transcript']
                    ]
test = subset.loc[
    subset.split=="test",  ['hashname',
                            'human_transcript',
                            'unnormalized_transcript']
                    ]
dev = subset.loc[
    subset.split=="dev",  ['hashname',
                            'human_transcript',
                            'unnormalized_transcript']
                    ]

train.to_csv("17_train.csv", index=False)
dev.to_csv("17_dev.csv", index=False)
test.to_csv("17_test.csv", index=False)

# Implementing the unnormalized preprocessing:

import parse
s = df.unnormalized_transcript[1]
s

p = parse.compile("{hit:d}.")
for seg in s.split():
    result = p.parse(seg)
    if result:
        print(result)

def process(text:str):
    from parse import compile
    from string import punctuation
    p = compile("{hit:d}.")
    in_list = text.split()
    out_list = list()
    for seg in in_list:
        parse_result = p.parse(seg)
        if parse_result:
            # We got a number with a dot afterward:
            out_list.append(seg.lower())
        else:
            out_list.append(seg.translate(str.maketrans('', '', punctuation)).lower())
    return " ".join(out_list)



train['unnormalized_transcript'] = train.unnormalized_transcript.apply(process)
test['unnormalized_transcript'] = test.unnormalized_transcript.apply(process)
dev['unnormalized_transcript'] = dev.unnormalized_transcript.apply(process)
train.to_csv("17_train.csv", index=False)
dev.to_csv("17_dev.csv", index=False)
test.to_csv("17_test.csv", index=False)


df.human_transcript[0]

pd.set_option('display.max_colwidth', None)
test.unnormalized_transcript.head(20)
