==> 22_paper_related.ipynb <==
import pandas as pd

train = pd.read_csv("17_train.csv")
test = pd.read_csv("17_test.csv")
dev = pd.read_csv("17_dev.csv")

train.tail()

24500*20/3600

rez = pd.read_csv("21_results.csv")
rez["wer"] = rez.wer.apply(lambda x: f"{100*x:0.2f}\%")
rez["cer"] = rez.cer.apply(lambda x: f"{100*x:0.2f}\%")
c1 = rez.modelname.str.startswith("19")
c2 = rez.steps == 3028

rez.loc[c1&c2, :]


from datasets import load_metric
wer_metric = load_metric("wer")
cer_metric = load_metric("cer")

cer_metric.compute(predictions = ["Hello", "Palestine"],
           references = ["Hello",  "Ponce de Leon"]
        )

import datasets

datasets.__version__

import this
