==> 23_published_model_performance_on_our_test_data.ipynb <==
from datasets import load_metric
wer_metric = load_metric("wer")
cer_metric = load_metric("cer")

import pandas as pd

test_df = pd.read_csv("17_test.csv")

from transformers import Wav2Vec2Processor, Wav2Vec2ForCTC
from datasets import Audio
import soundfile as sf
import torch
import os

# load model and tokenizer
processor = Wav2Vec2Processor.from_pretrained(
    "classla/wav2vec2-xls-r-parlaspeech-hr")
model = Wav2Vec2ForCTC.from_pretrained("classla/wav2vec2-xls-r-parlaspeech-hr")

test_df["audio"] = test_df.hashname.apply(lambda s: "transfer_10/"+s)
import datasets
from datasets import load_dataset, load_metric, Audio



def get_transcript(audio_filepath:str):
    device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
    speech, sample_rate = sf.read(audio_filepath)
    input_values = processor(speech, sampling_rate=sample_rate, return_tensors="pt").input_values.cuda()

    logits = model.to(device)(input_values.to(device)).logits

    predicted_ids = torch.argmax(logits, dim=-1)
    transcription = processor.decode(predicted_ids[0]).lower()
    return transcription
test_df["predicted"] = test_df.audio.apply(get_transcript)

audio_filepath=test_df.audio[0]

device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
speech, sample_rate = sf.read(audio_filepath)
input_values = processor(speech, sampling_rate=sample_rate, return_tensors="pt").input_values.to(device)

logits = model.to(device)(input_values).logits

predicted_ids = torch.argmax(logits, dim=-1)
transcription = processor.decode(predicted_ids[0]).lower()
transcription

cer = cer_metric.compute(references = test_df.human_transcript,
                 predictions = test_df.predicted
                )
cer

wer = wer_metric.compute(references = test_df.human_transcript,
                 predictions = test_df.predicted
                )
wer
