==> 29_final_data_preparation.ipynb <==
import pandas as pd

df = pd.read_json("ParlaSpeech-HR.v1.0.jsonl", orient="records", lines=True)

dev = df[df.split=="dev"]
test = df[df.split=="test"]

needed_files = dev.path.values.tolist() + test.path.values.tolist()
needed_files = set(needed_files)

Ok, I shall transfer these files to kt-gpu-vm-1Tb

for file in needed_files:
    os.system(
        f'cp data_16000_mono/{file} transfer_10/{file}' 
    )

# 'Full scale' data preparation:

I shall prepare a dataset just big enough to fit in remaining disk space on kt-gpu-cm-1TB.

import pandas as pd

df = pd.read_json("ParlaSpeech-HR.v1.0.jsonl", orient="records", lines=True)

import os
def cp(file):
    if os.path.exists(os.path.join(
                                    "/home/rupnik/macocu/task8/transfer_10",
                                    file
                                    )):
        return
    os.system(
        f'cp data_16000_mono/{file} transfer_10/{file}' 
    )
from concurrent.futures import ProcessPoolExecutor
with ProcessPoolExecutor(max_workers=60) as executor:
    executor.map(cp, df.path)
