==> 30_eval_on_ParlaSpeech-HR.v1.0.jsonl.ipynb <==
def process(text:str):
    from parse import compile
    from string import punctuation
    p = compile("{hit:d}.")
    in_list = text.split()
    out_list = list()
    for seg in in_list:
        parse_result = p.parse(seg)
        if parse_result:
            # We got a number with a dot afterward:
            out_list.append(seg.lower())
        else:
            out_list.append(seg.translate(str.maketrans('', '', punctuation)).lower())
    return " ".join(out_list)


import os
def correct_path(s):
    basename = os.path.basename(s)
    current_dir = "/home/peterr/macocu/task8/transfer_10/"
    return os.path.join(current_dir, basename)
import pandas as pd
df = pd.read_json("ParlaSpeech-HR.v1.0.jsonl", orient="records", lines=True)

df["path"] = df.path.apply(correct_path)
df["norm_text"] = df.norm_words.apply(" ".join)


dev = df[df.split=="dev"]
test = df[df.split=="test"]


dev.head(2)

from datasets import load_metric
wer_metric = load_metric("wer")
cer_metric = load_metric("cer")

import pandas as pd
import os
import torch
from transformers import Wav2Vec2Processor, Wav2Vec2ForCTC
from transformers import Wav2Vec2FeatureExtractor
from transformers import Wav2Vec2Processor, Wav2Vec2CTCTokenizer
import soundfile as sf
import os
# Use old or new vocab? 
os.system("cp vocab_300.json vocab.json")



tokenizer = Wav2Vec2CTCTokenizer.from_pretrained(
        "/home/peterr/macocu/task8/", unk_token="[UNK]", 
        pad_token="[PAD]", word_delimiter_token=" ")
feature_extractor = Wav2Vec2FeatureExtractor(
    feature_size=1, sampling_rate=16000, padding_value=0.0, do_normalize=True, return_attention_mask=True)

processor = Wav2Vec2Processor(
    feature_extractor=feature_extractor, tokenizer=tokenizer)
model = Wav2Vec2ForCTC.from_pretrained("/home/peterr/macocu/task8/25_normalised_200h/checkpoint-2720").cuda()

def transcribe(path):
    speech, sample_rate = sf.read(path)
    input_values = processor(speech, sampling_rate=sample_rate, return_tensors="pt").input_values.cuda()

    logits = model(input_values).logits

    predicted_ids = torch.argmax(logits, dim=-1)
    transcription = processor.decode(predicted_ids[0]).lower()
    return transcription



for df in [test, dev]:
    df["predicted"] = df.path.apply(transcribe)

for name, df in zip(["dev", "test"], [dev, test]):
    cer = cer_metric.compute(
        references = df.norm_text,
        predictions = df.predicted
    )
    wer = wer_metric.compute(
        references = df.norm_text,
        predictions = df.predicted
    )

    print(f"split: {name}, {cer=:0.4f}, {wer=:0.4f}")

dev.to_csv("30_eval_results_dev.csv", index=False)
test.to_csv("30_eval_results_test.csv", index=False)

# Per speaker wer and cer calculation

df = pd.concat([df, df.speaker_info.apply(pd.Series)], axis=1)
df.head(2)

results = []
for split in ("test", "dev"):
    subset = df[df.split == split]
    for speaker in subset.Speaker_name.unique():
        subsubset = subset[subset.Speaker_name == speaker].copy()
        subsubset["predicted"] = subsubset.path.apply(transcribe)
        speaker_count = subsubset.shape[0]
        cer = cer_metric.compute(
        references = subsubset.norm_text,
        predictions = subsubset.predicted
        )
        wer = wer_metric.compute(
        references = subsubset.norm_text,
        predictions = subsubset.predicted
        )

        print(f"{split=}, {speaker=}, {cer=:0.4f}, {wer=:0.4f}")
        results.append(dict(split=split, speaker=speaker, cer=cer, wer=wer, count=speaker_count))
        

per_speaker_df = pd.DataFrame(data=results).sort_values(by="cer", ascending=False)
per_speaker_df

per_speaker_df.groupby("split").mean()

df[df.split=="train"].shape[0]

free = 183
transfer_10_now = 23
all_samples = 214

i_need = all_samples - transfer_10_now
i_need
