==> 33_renaming_files.ipynb <==
import os
d = "data_16000_mono"

all_wavs = os.listdir(d)
all_paths = [os.path.join(
    d,
    file
) for file in all_wavs]

def rename(file):
    os.system(f"mv {os.path.join(d, file)} {os.path.join(d, 'seg.'+file)}")

from concurrent.futures import ProcessPoolExecutor
with ProcessPoolExecutor(max_workers=60) as executor:
    executor.map(rename, all_wavs)
