==> 34__dataset_validation.ipynb <==
import pandas as pd

df = pd.read_json("ParlaSpeech-HR.v1.0.jsonl", orient="records", lines=True)
df["sentence"] = df.norm_words.apply(" ".join)
df = df.rename(columns={"path":"hashname"})


from pydub import AudioSegment
import os
def test(hashname):
    fullpath = f"/home/rupnik/macocu/task8/data_16000_mono/seg.{hashname}"
    try:
        audio = AudioSegment.from_wav(fullpath)
        assert len(audio) != 0
    except:
        return fullpath


from concurrent.futures import ProcessPoolExecutor, as_completed
with ProcessPoolExecutor(max_workers=60) as executor:
    futures = [executor.submit(test, hashname) for hashname in df.hashname]
    for future in as_completed(futures):
        if future.result():
            print("Found faulty:", future.result())
