==> 37_dataprep_300h.ipynb <==
import pandas as pd


df = pd.read_json("ParlaSpeech-HR.v1.0.jsonl", orient="records", lines=True)
df = df.rename(columns={"path":"hashname"})

def process(text:str):
    from parse import compile
    from string import punctuation
    p = compile("{hit:d}.")
    in_list = text.split()
    out_list = list()
    for seg in in_list:
        parse_result = p.parse(seg)
        if parse_result:
            # We got a number with a dot afterward:
            out_list.append(seg.lower())
        else:
            out_list.append(seg.translate(str.maketrans('', '', punctuation)).lower())
    return " ".join(out_list)
df["sentence"] = df.words.apply(" ".join).apply(process)


df = pd.concat([df, df.speaker_info.apply(pd.Series)], axis=1)

df["path"] = df.hashname.apply(lambda s: "transfer/seg."+s)

test_df = df[df.split=="dev"].copy()

test_df.head(2)

datadir = "/home/rupnik/macocu/task8/data_16000_mono"
transferdir = "/home/rupnik/macocu/task8/transfer"
import os
import shutil
for name in test_df.hashname.values:
    shutil.copy(
        os.path.join(
            datadir, "seg."+name
        ),
        os.path.join(transferdir, name)
        
        )

test_df.to_csv("37_test_split.csv", index=False)

train_df = df[df.split=="train"].copy()

# Sampling train dataset:

same protocol as before. Add instance from every speaker until we get to 300h.

target_count = 300 * 3600 / 20
target_count


indices_to_keep = []


while len(indices_to_keep) < target_count:
    print(len(indices_to_keep), end="\r")
    current_slice = train_df.loc[~train_df.index.isin(indices_to_keep), :]
    for speaker in current_slice.Speaker_name.unique():
        index = current_slice.loc[current_slice.Speaker_name == speaker, :].index[0]
        indices_to_keep.append(index)

train_df = train_df.loc[train_df.index.isin(indices_to_keep), :].copy()

datadir = "/home/rupnik/macocu/task8/data_16000_mono"
transferdir = "/home/rupnik/macocu/task8/transfer"
import os
import shutil
for name in train_df.hashname.values:
    shutil.copy(
        os.path.join(
            datadir, "seg."+name
        ),
        os.path.join(transferdir, name)
        
        )

train_df.to_csv("37_train_split.csv", index=False)

train_df.shape
