==> 39_1_adding_LM.ipynb <==
modelname = "38_unnormalised/checkpoint-6700/"

from transformers import Wav2Vec2Processor, Wav2Vec2ForCTC

processor = Wav2Vec2Processor.from_pretrained(modelname)
model = Wav2Vec2ForCTC.from_pretrained(modelname)

# with open("5gram.arpa", "r") as read_file, open("5gram_correct.arpa", "w") as write_file:
#   has_added_eos = False
#   from tqdm.auto import tqdm
#   for line in tqdm(read_file):
#     if not has_added_eos and "ngram 1=" in line:
#       count=line.strip().split("=")[-1]
#       write_file.write(line.replace(f"{count}", f"{int(count)+1}"))
#     elif not has_added_eos and "<s>" in line:
#       write_file.write(line)
#       write_file.write(line.replace("<s>", "</s>"))
#       has_added_eos = True
#     else:
#       write_file.write(line)

from transformers import AutoProcessor

processor = AutoProcessor.from_pretrained(modelname,    eos_token="</s>", bos_token="<s>")

vocab_dict = processor.tokenizer.get_vocab()
sorted_vocab_dict = {k: v for k, v in sorted(vocab_dict.items(), key=lambda item: item[1]) if k != "<pad>"}

sorted_vocab_dict.keys()

from pyctcdecode import build_ctcdecoder

decoder = build_ctcdecoder(
    labels=list(sorted_vocab_dict.keys()),
    kenlm_model_path="5gram.bin",
)

from transformers import Wav2Vec2ProcessorWithLM

processor_with_lm = Wav2Vec2ProcessorWithLM(
    feature_extractor=processor.feature_extractor,
    tokenizer=processor.tokenizer,
    decoder=decoder,
)

processor_with_lm.save_pretrained("38_with_lm/wav2vec2-xls-r-parlaspeech-hr",
)

model.save_pretrained("38_with_lm/wav2vec2-xls-r-parlaspeech-hr",)

# Reevaluating

import pandas as pd
import numpy as np
import os

transferdir = "transfer/"

def process(text: str):
    from parse import compile
    from string import punctuation

    p = compile("{hit:d}.")
    in_list = text.split()
    out_list = list()
    for seg in in_list:
        parse_result = p.parse(seg)
        if parse_result:
            # We got a number with a dot afterward:
            out_list.append(seg.lower())
        else:
            out_list.append(seg.translate(str.maketrans("", "", punctuation)).lower())
    return " ".join(out_list)

df = pd.read_json("ParlaSpeech-HR.v1.0.jsonl", orient="records", lines=True)
df = df.rename(columns={"path":"hashname"})
df = df.loc[df.split.isin("test,dev".split(",")), :]

df["sentence"] = df.words.apply(" ".join).apply(process)
df["path"] = df.hashname.apply(lambda s: os.path.join(transferdir, s))

from tqdm.auto import tqdm


from itertools import groupby
import torch
from transformers import Wav2Vec2Processor, Wav2Vec2ForCTC
from transformers import Wav2Vec2FeatureExtractor
from transformers import Wav2Vec2Processor, Wav2Vec2CTCTokenizer
import soundfile as sf

import os

model = Wav2Vec2ForCTC.from_pretrained(modelname).cuda()

def get_transcript(audio_filepath:str):
    speech, sample_rate = sf.read(audio_filepath)
    input_values = processor_with_lm(speech, sampling_rate=sample_rate, return_tensors="pt").input_values.cuda()

    logits = model(input_values).logits

    predicted_ids = torch.argmax(logits, dim=-1)
    transcription = processor_with_lm.decode(predicted_ids[0]).lower()


    words = [w for w in transcription.split(' ') if len(w) > 0]

    return " ".join(words)

from tqdm.auto import tqdm
df["predictions"] = [get_transcript(path) for path in tqdm(df.path.values)]


from datasets import load_metric
wer_metric = load_metric("wer")
cer_metric = load_metric("cer")

for splt in ["dev", "test"]:
    print("Evaluating on ", splt)
    wer = wer_metric.compute(
        references=df.loc[df.split==splt, "sentence"],
        predictions=df.loc[df.split==splt, "predictions"]
    )

    cer = cer_metric.compute(
        references=df.loc[df.split==splt, "sentence"],
        predictions=df.loc[df.split==splt, "predictions"]
    )

    print(f"{wer=:0.4f}, {cer=:0.4f}")

processor.tokenizer.get_vocab()

audio_filepath = df.path.values[2]
speech, sample_rate = sf.read(audio_filepath)
input_values = processor_with_lm(speech, sampling_rate=sample_rate, return_tensors="pt").input_values.cuda()

logits = model(input_values).logits

predicted_ids = torch.argmax(logits, dim=-1)
transcription = processor_with_lm.batch_decode(predicted_ids[0])
transcription


input_values = processor(speech, sampling_rate=sample_rate, return_tensors="pt").input_values.cuda()

logits = model(input_values).logits
predicted_ids = torch.argmax(logits, dim=-1)
transcription = processor.decode(predicted_ids[0]).lower()
logits.shape

input_values = processor_with_lm(speech, sampling_rate=sample_rate, return_tensors="pt").input_values.cuda()

logits = model(input_values).logits
predicted_ids = torch.argmax(logits, dim=-1)
#transcription = processor_with_lm.decode(predicted_ids[0]).lower()
logits

predicted_ids[0]
